Add Type enforce for canonicalURL param, and return statuses for getArticleIDFromCanonicalURL

+++ src/Facebook/InstantArticles/Client/Client.php
@@ -100,12 +100,22 @@ class Client
     /**
      * Removes an article from your Instant Articles library.
      *
-     * @param string $canonicalURL The canonical URL of the article to delete.
+     * @param $canonicalURL
+     * @return \Facebook\InstantArticles\Client\InstantArticleStatus
      */
-    public function removeArticle($canonicalURL) {
+    public function removeArticle($canonicalURL)
+    {
+        if (!$canonicalURL) {
+            return InstantArticleStatus::notFound(array('$canonicalURL param not passed to ' . __FUNCTION__ . '.'));
+        }
+
+        Type::enforce($canonicalURL, Type::STRING);
+
         if ($articleID = $this->getArticleIDFromCanonicalURL($canonicalURL)) {
             $this->facebook->delete($articleID);
+            return InstantArticleStatus::success(array('An Instant Article ID ' . $articleID . ' was found for ' . $canonicalURL . ' in ' . __FUNCTION__ . '.'));
         }
+        return InstantArticleStatus::notFound(array('An Instant Article ID ' . $articleID . ' was not found for ' . $canonicalURL . ' in ' . __FUNCTION__ . '.'));
     }
 
     /**
